+++ CHANGELOG.rst
@@ -15,9 +15,11 @@ Breaking changes
 
 Deprecations
 -------------
+- ``result.fminfull`` is deprecated and will be removed in the future. Use ``result.fmin`` instead.
 
 Bug fixes and small changes
 ---------------------------
+- ``result.fmin`` now returns the full likelihood, while ``result.fminopt`` returns the optimized likelihood with potential constant subtraction. The latter is mostly used by the minimizer and other libraries. This behavior is consistent with the behavior of other methods in the loss that return by default the full, unoptimized value.
 
 Experimental
 ------------